[refactor] Adapter 클래스 위치를 adapter에서 꺼냄
