+++ src/main/java/kr/co/pawong/pwbe/infrastructure/ai/adapter/huggingface/HuggingFaceEmbeddingModel.java
@@ -1,4 +1,4 @@
-package kr.co.pawong.pwbe.infrastructure.ai.huggingface;
+package kr.co.pawong.pwbe.infrastructure.ai.adapter.huggingface;
 
 import java.util.ArrayList;
 import java.util.List;
