+++ src/main/java/kr/co/pawong/pwbe/infrastructure/ai/adapter/openai/OpenAiChatAdapter.java
@@ -1,4 +1,4 @@
-package kr.co.pawong.pwbe.infrastructure.ai.openai;
+package kr.co.pawong.pwbe.infrastructure.ai.adapter.openai;
 
 import static kr.co.pawong.pwbe.infrastructure.ai.enums.RefineMsg.REFINE_TEMPLATE_1;
 import static kr.co.pawong.pwbe.infrastructure.ai.enums.TaggingMsg.TAGGING_TEMPLATE_1;
